fix(frontend): refresh folder counts and handle rejection on file delete
Move the file menu's state and handlers into useFileActionsMenu. On a
successful delete also invalidate the folder list and storage usage so
folder file counts and the footer don't go stale, and catch the
mutateAsync rejection (the toast already reports it).
